+++ nodeconductor/structure/tests/unittests/test_registry.py
@@ -2,7 +2,7 @@ from django.test import TestCase
 
 from nodeconductor.structure import SupportedServices, ServiceBackendNotImplemented
 from nodeconductor.structure.tests import TestConfig, TestBackend
-from nodeconductor.structure.tests.models import TestService, TestInstance, TestNewInstance
+from nodeconductor.structure.tests.models import TestService, TestNewInstance
 from nodeconductor.structure.tests.serializers import ServiceSerializer
 
 
@@ -20,9 +20,9 @@ class ServiceRegistryTest(TestCase):
                          SupportedServices.get_service_serializer(TestService))
 
     def test_get_service_resources(self):
-        self.assertEqual([TestInstance, TestNewInstance],
+        self.assertEqual([TestNewInstance],
                          SupportedServices.get_service_resources(TestService))
 
     def test_model_key(self):
         self.assertEqual(TestConfig.service_name,
-                         SupportedServices.get_model_key(TestInstance))
+                         SupportedServices.get_model_key(TestNewInstance))